Implement active index loading.
The search tab manager will activate the index at initial time when
opening (and switching) an ontology. If the index is present and the
active ontology has the same index checksum then the search manager will
reuse the index, or otherwise it will build the index and save it for
future reuse.