Support for full string matching
For starts with and exact match, added a new version that searches the entire string rather than
the tokenized version. This only applies to search over rdfs:label and is intended to  support a common
use case where the users wishes to find the class Cancer or Gene and does not want to find all the classes
where the word occurs in the label.